Improve safety of tcb name access
The boot code now leverages invokeUntyped_Retype to create its kernel
objects, including thread TCBs. This means that there will be two
threads without parents. Names of threads with no parents must be
manually configured and this is done for both the idle and initial
threads in the boot code.