smp: Don't check for IPIs in sel4_atomic_exchange
The CLH lock is a FIFO lock and the sel4_atomic_exchange operation is
only used for adding the node to the end of the lock queue. Even if all
nodes arrive at the lock at the same time, the last node to succeed will
only wait for as long as it takes for other nodes to complete the
exchange first.