gic_v3: Add special LPI handling
LPI are a 4th kind of interrupt for "message signalled interrupts". If
they are supported by a GICv3 implementation then they may be coupled
with an interrupt translation service(ITS). With this set of changes,
the seL4 irq handling model can still support physical LPI interrupts
provided that the LPI redistributor and ITS service are correctly
managed by another system component. A safe way to do this is to
statically configure before the kernel boots and the lack of untypeds to
the resources used can prevent any other access during runtime.